Add dbupdate script to Dockerfile
dbupdate.sh is a script that will add columns to the database. Useful so
you don't have to re-migrate your entire database when there are
changes. Unfortunately sqlite only allows adding and renaming columns,
so anything more than adding a column would require a re-migration.

Postgres is the obvious solution here, but I would really like to keep
this application to a single container, and easy to setup for people not
familiar with containers. If this ever grows to be a very large product
then this will be unavoidable, but we aren't there yet :)